refactor(routes): extract /agent-templates into agent_templates.js (god-file decomposition)
Route contract unchanged (284 routes), 0 /agent-templates left inline, vitest 38 passed 268 passed . Lucy (GLM-5.2), driver-gated.
